ECE 568/Ride Sharing Web-App/docker-deploy/web-app/requirements.txt
Django>=2.0,<3.0
psycopg2
django-bootstrap-form
django-crispy-forms
django-filter
